Make `detail::NumProperties::size` a signed type
Since `detai::NumProperties::size` is normally used in comparisons with
other signed types, make it signed as well, to prevent having to
unsafely cast a signed type for which we usually know it will be
positive (which we do not check, hence the unsafety) to an unsigned
type.

+++ include/plywoot/plywoot/writer.hpp
@@ -341,7 +341,7 @@ private:
 
     // In case the element defines more properties than the source data,
     // append the missing properties with a default value of zero.
-    if (detail::numProperties<Ts...>() < static_cast<std::size_t>(std::distance(first, last)))
+    if (detail::numProperties<Ts...>() < std::distance(first, last))
     {
       this->writeMissingProperties(first + detail::numProperties<Ts...>(), last);
     }
